refactor: useAutoScrollIntoView を useBookSuggestions に統合
スクロール制御のロジックを useBookSuggestions に吸収し、
useAutoScrollIntoView を削除してファイル数を削減した。
